dockerizziamo/requirements.txt
django
djangorestframework
django-crispy-forms
crispy-bootstrap4
requests